Product service

+++ src/controllers/scheduleControl.js
@@ -4,9 +4,8 @@ const bodyParser=require('body-parser');
 const cors = require('cors');
 const { Op } = require("sequelize");
 
-require("dotenv-safe").config();
-
 const Schedule=require('../models/Schedule');
+const ProSchedule=require('../models/ManyToMany_Models/ProductSchedule');
 const AdjustTime=require('../middleware/AdjustTime');
 
 router.use(cors())
@@ -14,15 +13,23 @@ router.use(cors())
 router.post('/register',AdjustTime,async(req,res)=>{
 	try{
 		const result=await Schedule.create({
-			SeviceProvided:req.body.SeviceProvided,
-			Description:req.body.Description,
 			ScheduledDay:req.body.ScheduledDay,
 			ScheduledHour:req.body.ScheduledHour,
 			ClientName:req.body.ClientName,
 			userId:req.body.userId
 		})
-		if(result){
-			res.json({menssage:"Sucess on Create"})
+		if(result){	
+			for(var x=0;x<req.body.ProSerId.length;x++){
+				const resp=await ProSchedule.create({
+					ScheduleId:result.id,
+					ProSerId:req.body.ProSerId[x]
+				})
+				if(!resp){
+					res.json({error:"Could not Create"});
+				}
+			}
+			
+			res.json({menssage:"Success on Create"});
 		}
 	}catch(err){
 		res.status(400).send({Error:"Creation Failed"});
@@ -39,7 +46,21 @@ router.get('/getOne',async(req,res)=>{
 		});
 		if(result){
 			const Data=result.dataValues;
-			res.json(Data)
+			const resp=await ProSchedule.findAll({
+				where:{
+					ScheduleId:Data.id
+				}
+			})
+			if(resp){
+				const Data2=resp.map(function(item,ID){
+					let ScheduleId=item.ScheduleId,
+					ProSerId=item.ProSerId
+
+					return {ScheduleId,ProSerId}
+				});
+				res.json(Object.assign(Data,Data2));
+			}
+			
 		}else{
 			res.json({menssage:"Cannot Find any register at this time"})
 		}
@@ -57,28 +78,64 @@ router.get('/getAllFromDay',async(req,res)=>{
 			raw:true
 		})
 		if(result){
-			const data=result.map(function(item,id){
-				return item;
+			const Data=result.map(function(item,ID){
+				let id=item.id,
+				ScheduledDay=item.ScheduledDay,
+				ScheduledHour=item.ScheduledHour,
+				ClientName=item.ClientName
+				return {id,ScheduledDay, ScheduledHour, ClientName};
 			})
-			res.json(data)
+
+			var obj={};
+			for(var x=0;x<Data.length;x++){
+				const resp=await ProSchedule.findAll({
+					where:{
+						ScheduleId:Data[x].id
+					}
+				});
+				if(resp){
+					const Data2=resp.map(function(item,ID){
+						let ScheduleId=item.ScheduleId,
+						ProSerId=item.ProSerId
+						
+						return {ScheduleId,ProSerId}
+					});
+
+					obj[x]=Data2;
+				}
+			}
+
+			var end = Data.concat(obj);
+
+			res.json(end);
 		}
-	}catch{
+	}catch(err){
 		res.status(400).send({Error:"Error"});
 	}
 })
 
 router.delete('/delete/One',async(req,res)=>{
 	try{
-		const del=await Schedule.destroy({
+		const find=await Schedule.findOne({
 			where:{
 				[Op.and]:[{ScheduledDay:req.body.ScheduledDay },{ScheduledHour:req.body.ScheduledHour },{userId:req.body.userId}]
 			}
 		})
-		if(del){
-			res.json({menssage:"Shedule deleted"})
-		}
-		else{
-			res.json({Error:"Schedule not deleted"})
+		if(find){
+			const DelX=await ProSchedule.destroy({
+				where:{ScheduleId:find.id}
+			})
+			if(DelX){
+				const del=await Schedule.destroy({
+					where:{id:find.id}
+				})
+				if(del){
+					res.json({menssage:"Shedule deleted"})
+				}
+				else{
+					res.json({Error:"Schedule not deleted"})
+				}
+			}
 		}
 	}catch(err){
 		console.log(err)
@@ -86,25 +143,41 @@ router.delete('/delete/One',async(req,res)=>{
 	}
 })
 
-router.delete('/delete/Day',async(req,res)=>{
+router.delete('/delete/EntireDay',async(req,res)=>{
 	try{
-		const del=await Schedule.destroy({
+		const find=await Schedule.findAll({
 			where:{
 				[Op.and]:[{ScheduledDay:req.body.ScheduledDay }, { userId:req.body.userId }]
-			}
+			},raw:true
 		});
-		if(del){
-			res.json({menssage:"Records Deleted Successfully"});
-		}
-		else{
-			res.json({menssage:"Records not Deleted"});
+
+		if(find){
+			const data=await find.map(function(item,ID){
+				return item.id
+			})
+			for(var x=0;x<data.length;x++){
+				const delx=await ProSchedule.destroy({
+					where:{ScheduleId:data[x]}
+				})
+			}
+			const del=await Schedule.destroy({
+				where:{
+					[Op.and]:[{ScheduledDay:req.body.ScheduledDay }, { userId:req.body.userId }]
+				}
+			});
+			if(del){
+				res.json({menssage:"Records Deleted Successfully"});
+			}
+			else{
+				res.json({menssage:"Records not Deleted"});
+			}
 		}
 	}catch(err){
 		res.status(400).send({Error:"Error"});
 	}
 })
 
-router.put('/update/Day',AdjustTime,async(req,res)=>{
+router.put('/update',AdjustTime,async(req,res)=>{
 	try{
 		const result=await Schedule.findOne({
 			where:{
@@ -112,7 +185,7 @@ router.put('/update/Day',AdjustTime,async(req,res)=>{
 			}
 		});
 		if(result){
-			const Types='{"type":["SeviceProvided","Description","ScheduledDay","ScheduledHour","ClientName"]}';
+			const Types='{"type":["ScheduledDay","ScheduledHour","ClientName"]}';
 			const obj=JSON.parse(Types);
 
 			for(var x=0;x<obj.type.length;x++){
